requirements.txt
# Core Data Science
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Google Cloud
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
db-dtypes==1.1.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Machine Learning
scikit-learn==1.3.2

# Jupyter
jupyter==1.0.0
jupyterlab==4.0.9
ipywidgets==8.1.1

# Data Quality
great-expectations==0.18.5

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Linting & Formatting
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14